+++ src/main/java/zenpad/ui/TabManager.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.swing.JTabbedPane;
 
 import zenpad.controllers.TabController;
+import zenpad.editor.EditorTab;
 import zenpad.views.TabView;
 
 public class TabManager {
